# j-enterprise-service-bus/tmp/openapi-gen-test/src/gen/java/my/service/PetsApi.java
package my.service;

import my.service.Error;
import my.service.Pet;

import java.util.List;
import javax.ws.rs.*;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiResponses;
import io.swagger.annotations.ApiResponse;

import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * Swagger Petstore
 *
 * <p>No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 */
@Path("/pets")
@Api(value = "/", description = "")
public interface PetsApi  {

    /**
     * Create a pet
     *
     */
    @POST
    
    @Consumes({ "application/json" })
    @Produces({ "application/json" })
    @ApiOperation(value = "Create a pet", tags={ "pets" })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Null response"),
        @ApiResponse(code = 200, message = "unexpected error", response = Error.class) })
    public void createPets(@Valid @NotNull Pet pet);

    /**
     * List all pets
     *
     */
    @GET
    
    @Produces({ "application/json" })
    @ApiOperation(value = "List all pets", tags={ "pets" })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "A paged array of pets", response = Pet.class, responseContainer = "List"),
        @ApiResponse(code = 200, message = "unexpected error", response = Error.class) })
    public List<Pet> listPets(@QueryParam("limit") @Max(100) Integer limit);

    /**
     * Info for a specific pet
     *
     */
    @GET
    @Path("/{petId}")
    @Produces({ "application/json" })
    @ApiOperation(value = "Info for a specific pet", tags={ "pets" })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Expected response to a valid request", response = Pet.class),
        @ApiResponse(code = 200, message = "unexpected error", response = Error.class) })
    public Pet showPetById(@PathParam("petId") String petId);
}